Histexprt init
Meant to read the given data file and export it.

Currently reading the test data, need to find a way to write it. Once done so, we can comment out the "PERFORM P320-START THRU P320-END" and switch for P330 instead, as P320 is for displaying the data;

Also, instead of reading the character limit, checking if tipreg is equal to 01, as it is the only one we want to read for this case.